refactor: judgeWebApp 기능을 Service로 분리
