+++ src/controllers/judgeController.ts
@@ -1,24 +1,11 @@
-import fs from "fs";
-
-import * as chromeLauncher from "chrome-launcher";
 import { Request, Response } from "express";
 
-export const judgeWebApp = async (req: Request, res: Response) => {
-  const chrome = await chromeLauncher.launch({ chromeFlags: ["--headless"] });
-  const lighthouse = (await import("lighthouse")).default;
-  
-  const runnerResult = await lighthouse(`http://localhost:3000/${req.id}.html`, {
-    logLevel: "info",
-    port: chrome.port,
-  }, {
-    extends: "lighthouse:default",
-    settings: {
-      locale: "ko",
-      onlyCategories: ["accessibility"],
-    }
-  });
+import * as JudgeService from "@/services/judgeService";
 
-  if (!runnerResult) {
+export const judgeWebApp = async (req: Request, res: Response) => {
+  const result = await JudgeService.judgeWebApp(`http://localhost:3000/${req.id}.html`);
+
+  if (!result) {
     return res.status(500).json({
       result: "fail",
     })
@@ -26,6 +13,6 @@ export const judgeWebApp = async (req: Request, res: Response) => {
 
   return res.status(200).json({
     result: "sucess",
-    report: JSON.stringify(runnerResult.lhr),
+    report: JSON.stringify(result.lhr),
   })
 };
